requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
python-dotenv==1.0.0
yt-dlp
pymongo
motor
dnspython==2.4.2 